requirements.txt
ffmpeg-python~=0.2.0
music-tag~=0.4.3
